chore: rename webgl-object to mesh
